sqlstr.Base.build implementation
sqlstr.Base.build is an instance method that will be used to build
valid sql strings from template strings. It has been changed from
a class method to an instance method so as to avoid sharing of
context by all instances of the class.

Some minor change in sqlstr/SQLite3.py